Refac: 기존 홈화면 동아리 리스트 api ClubController에 통합
