+++ src/main/java/com/cluting/clutingbackend/plan/controller/ClubController.java
@@ -4,6 +4,9 @@ import com.cluting.clutingbackend.plan.dto.request.ClubCreateRequestDto;
 import com.cluting.clutingbackend.plan.dto.response.ClubResponseDto;
 import com.cluting.clutingbackend.plan.service.ClubService;
 import com.cluting.clutingbackend.util.security.CustomUserDetails;
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.responses.ApiResponse;
+import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
@@ -18,14 +21,42 @@ import java.util.List;
 public class ClubController {
     private final ClubService clubService;
 
-    // 가장 인기 있는 동아리
+    @Operation(description = "동아리 id로 조회")
+    @ApiResponses({
+            @ApiResponse(responseCode = "200", description = "동아리 조회 성공"),
+            @ApiResponse(responseCode = "404", description = "Club not found"),
+            @ApiResponse(responseCode = "500", description = "Internal server error")})
+    @GetMapping("/{clubId}")
+    public ClubResponseDto findClub(@PathVariable("clubId") Long clubId) {
+        return clubService.findClub(clubId);
+    }
+
+    @Operation(description = "로그인 된 사용자가 가입한 동아리 목록 조회")
+    @ApiResponses({
+            @ApiResponse(responseCode = "200", description = "동아리 목록 조회 성공"),
+            @ApiResponse(responseCode = "404", description = "User not found"),
+            @ApiResponse(responseCode = "500", description = "Internal server error")})
+    @GetMapping("/userClubs")
+    public List<ClubResponseDto> findClubs(@AuthenticationPrincipal CustomUserDetails userDetails) {
+        return clubService.userClubs(userDetails.getUser());
+    }
+
+    @Operation(description = "가장 인기 있는 동아리 조회")
+    @ApiResponses({
+            @ApiResponse(responseCode = "200", description = "동아리 조회 성공"),
+            @ApiResponse(responseCode = "404", description = "Club not found"),
+            @ApiResponse(responseCode = "500", description = "Internal server error")})
     @GetMapping("/popular")
     @ResponseStatus(value = HttpStatus.OK)
     public List<ClubResponseDto> popular() {
         return clubService.popular();
     }
 
-    // 동아리 추가
+    @Operation(description = "동아리 추가")
+    @ApiResponses({
+            @ApiResponse(responseCode = "201", description = "동아리 추가 성공"),
+            @ApiResponse(responseCode = "404", description = "동아리 추가 실패"),
+            @ApiResponse(responseCode = "500", description = "Internal server error")})
     @PostMapping("/create")
     @ResponseStatus(value = HttpStatus.CREATED)
     public ClubResponseDto create(
@@ -35,7 +66,11 @@ public class ClubController {
         return clubService.create(userDetails.getUser(), clubCreateRequestDto, profile);
     }
 
-    // 리크루팅 시작
+    @Operation(description = "리크루팅 시작")
+    @ApiResponses({
+            @ApiResponse(responseCode = "200", description = "동아리 리크루팅 시작으로 상태 변경 성공"),
+            @ApiResponse(responseCode = "404", description = "Club not found"),
+            @ApiResponse(responseCode = "500", description = "Internal server error")})
     @PutMapping("/start-recruiting/{clubId}")
     @ResponseStatus(value = HttpStatus.OK)
     public ClubResponseDto startRecruiting(
