+++ src/main/java/com/cluting/clutingbackend/plan/service/ClubService.java
@@ -1,6 +1,7 @@
 package com.cluting.clutingbackend.plan.service;
 
 import com.cluting.clutingbackend.plan.domain.Club;
+import com.cluting.clutingbackend.plan.domain.ClubUser;
 import com.cluting.clutingbackend.plan.domain.User;
 import com.cluting.clutingbackend.plan.dto.request.ClubCreateRequestDto;
 import com.cluting.clutingbackend.plan.dto.response.ClubResponseDto;
@@ -19,6 +20,25 @@ import java.util.List;
 public class ClubService {
     private final ClubRepository clubRepository;
     private final AwsS3Service awsS3Service;
+
+    // 동아리 id로 조회
+    @Transactional(readOnly = true)
+    public ClubResponseDto findClub(Long clubId) {
+        return ClubResponseDto.toDto(
+                clubRepository.findById(clubId).orElseThrow(
+                        () -> new ResponseStatusException(
+                                HttpStatus.BAD_REQUEST, "존재하지 않는 동아리 입니다."
+                        )
+                )
+        );
+    }
+
+    // 로그인 된 사용자가 입한 동아리 목록
+    @Transactional(readOnly = true)
+    public List<ClubResponseDto> userClubs(User user) {
+        List<ClubUser> clubUsers = user.getClubUsers();
+        return clubUsers.stream().map(clubUser -> ClubResponseDto.toDto(clubUser.getClub())).toList();
+    }
 
     // 가장 인기 있는 동아리
     @Transactional(readOnly = true)
